update more function
