+++ client/src/App.tsx
@@ -6,11 +6,14 @@ import LoginRegister from "./Component/Pages/LoginRegister";
 import { BrowserRouter as Router, Switch, Route } from "react-router-dom";
 import Owe from "./Component/Pages/Owe";
 import ProtectedRoute from "./Protected/ProtectedRoute";
+import { getInforMoneyUser } from "./redux/actions/Money.action";
 
 const App: React.FC = () => {
   const dispatch = useDispatch();
   useEffect(() => {
     dispatch(getAllInforOfUser());
+    dispatch(getInforMoneyUser());
+
   }, [dispatch]);
   return (
     <Router>
