+++ client/src/redux/Types.tsx
@@ -1,12 +1,19 @@
-export const FETCH_DATA_USER = "FETCH_DATA";
-export const FETCH_DATA_USER_SUCCESS = "FETCH_DATA_SUCCESS";
-export const FETCH_DATA_USER_ERROR = "FETCH_DATA_ERROR";
-export const LOGIN_USER_REQUEST = "LOGIN_USER_REQUEST";
-export const LOGIN_USER_SUCCESS = "LOGIN_USER_SUCCESS";
-export const LOGIN_USER_ERROR = "LOGIN_USER_ERROR";
-export const REGISTER_USER_REQUEST = "REGISTER_USER_REQUEST";
-export const REGISTER_USER_SUCCESS = "REGISTER_USER_SUCCESS";
-export const REGISTER_USER_ERROR = "REGISTER_USER_ERROR";
-export const DRAW_USER_REQUEST = "DRAW_USER_REQUEST";
-export const DRAW_USER_SUCCESS = "DRAW_USER_SUCCESS";
-export const DRAW_USER_ERROR = "DRAW_USER_ERROR";
+export const FETCH_DATA_USER = "FETCH_DATA"
+export const FETCH_DATA_USER_SUCCESS = "FETCH_DATA_SUCCESS"
+export const FETCH_DATA_USER_ERROR = "FETCH_DATA_ERROR"
+export const LOGIN_USER_REQUEST = "LOGIN_USER_REQUEST"
+export const LOGIN_USER_SUCCESS = "LOGIN_USER_SUCCESS"
+export const LOGIN_USER_ERROR = "LOGIN_USER_ERROR"
+export const REGISTER_USER_REQUEST = "REGISTER_USER_REQUEST"
+export const REGISTER_USER_SUCCESS = "REGISTER_USER_SUCCESS"
+export const REGISTER_USER_ERROR = "REGISTER_USER_ERROR"
+export const DRAW_REQUEST = "DRAW_REQUEST"
+export const DRAW_SUCCESS = "DRAW_SUCCESS"
+export const DRAW_ERROR = "DRAW_ERROR"
+export const GET_INFOR_MONEY_USER_REQUEST = "GET_INFOR_MONEY_USER_REQUEST"
+export const GET_INFOR_MONEY_USER_SUCCESS = "GET_INFOR_MONEY_USER_SUCCESS"
+export const GET_INFOR_MONEY_USER_ERROR = "GET_INFOR_MONEY_USER_ERROR"
+export const DELETE_DRAW_REQUEST = "DELETE_DRAW_REQUEST"
+export const DELETE_DRAW_SUCCESS = "DELETE_DRAW_SUCCESS"
+export const DELETE_DRAW_ERROR = "DELETE_DRAW_ERROR"
+
